fix: Rainbow hue truncation in brightness selection
- Changed brightness_rainbow_hue from uint8_t to uint16_t
- Prevents truncation at 255 (hue range is 0-359)
- Fixes visible color jump at 252→258 transition
- Ensures smooth rainbow cycling throughout full spectrum

Addresses: CodeRabbit comment 2416826731

+++ src/common/lib/command_mode.c
@@ -105,7 +105,7 @@ static command_mode_context_t cmd_mode = {
 #ifdef CONVERTER_LEDS
 /** Brightness selection state tracking */
 static uint8_t brightness_original_value = 0;    /**< Original brightness when entering selection mode */
-static uint8_t brightness_rainbow_hue = 0;       /**< Current rainbow hue (0-359) for visual feedback */
+static uint16_t brightness_rainbow_hue = 0;      /**< Current rainbow hue (0-359) for visual feedback */
 #endif
 
 /** Command mode timing constants (milliseconds) */
